Update DlsDirectGrantX509Authenticator.java
To handle for users present and future that might lose their CAC attributes, I am adding setSingleAttribute for edipi, piv, and subjectdn to each login even for the CAC user. This is should add little to no overhead, and should the CAC values change, we would want to capture them anyway.